requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
gTTS==2.2.2
idna==2.10
pyTelegramBotAPI==3.7.6
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.3
